+++ src/main/java/com/aztsoft/gym/view/command/abstractcommand/Command.java
@@ -3,14 +3,12 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.aztsoft.gym.view;
+package com.aztsoft.gym.view.command.abstractcommand;
 
 /**
  *
- * @author windows
+ * @author arnold9108@gmail.com
  */
-public interface ViewForm {
-    
-    public void runView();
-    
+public interface Command {
+    void execute();
 }
