+++ src/main/java/com/aztsoft/gym/view/forms/form/CustomerCatalogForm.java
@@ -1,34 +1,23 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package com.aztsoft.gym.view;
+package com.aztsoft.gym.view.forms.form;
 
 import com.aztsoft.gym.controller.CustomerController;
 import com.aztsoft.gym.controller.CustomerControllerImp;
-import com.aztsoft.gym.view.command.Command;
-import com.aztsoft.gym.view.command.GetAllCustomerRecordsCommand;
+import com.aztsoft.gym.view.command.abstractcommand.Command;
+import com.aztsoft.gym.view.command.concrete.GetAllCustomerRecordsCommand;
+import com.aztsoft.gym.view.forms.abstractview.ViewForm;
+
 import javax.swing.table.TableModel;
 
 /**
  *
- * @author windows
+ * @author arnold9108@gmail.com
  */
 public final class CustomerCatalogForm extends javax.swing.JFrame implements ViewForm {
     
-    /**
-     * Creates new form CustomerCatalogForm
-     */
     public CustomerCatalogForm() {
         runView();
     }
 
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
     @SuppressWarnings("unchecked")
     // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
     private void initComponents() {
@@ -99,15 +88,7 @@ public final class CustomerCatalogForm extends javax.swing.JFrame implements Vie
         tblCustomerCatalog.setModel(model);
     }
     
-    /**
-     * @param args the command line arguments
-     */
     public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
         try {
             for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
                 if ("Nimbus".equals(info.getName())) {
@@ -118,8 +99,6 @@ public final class CustomerCatalogForm extends javax.swing.JFrame implements Vie
         } catch (ClassNotFoundException | InstantiationException | IllegalAccessException | javax.swing.UnsupportedLookAndFeelException ex) {
             java.util.logging.Logger.getLogger(CustomerCatalogForm.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         }
-
-        /* Create and display the form */
         java.awt.EventQueue.invokeLater(new Runnable() {
             @Override
             public void run() {
