+++ src/main/java/com/aztsoft/gym/view/forms/form/CustomerForm.java
@@ -1,16 +1,12 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package com.aztsoft.gym.view;
+package com.aztsoft.gym.view.forms.form;
 
 import com.aztsoft.gym.controller.CustomerController;
 import com.aztsoft.gym.controller.CustomerControllerImp;
 import com.aztsoft.gym.domain.Customer;
 import com.aztsoft.gym.domain.CustomerRegistration;
-import com.aztsoft.gym.view.command.AddCustomerCommand;
-import com.aztsoft.gym.view.command.Command;
+import com.aztsoft.gym.view.command.concrete.AddCustomerCommand;
+import com.aztsoft.gym.view.command.abstractcommand.Command;
+import com.aztsoft.gym.view.forms.abstractview.ViewForm;
 import com.toedter.calendar.JDateChooser;
 import java.awt.Image;
 import java.io.File;
@@ -28,24 +24,16 @@ import org.apache.commons.lang3.StringUtils;
 
 /**
  *
- * @author windows
+ * @author arnold9108@gmail.com
  */
 public final class CustomerForm extends javax.swing.JFrame implements ViewForm {
     private static final int VISIT = 0;
     public FileInputStream imageBlob = null;
 
-    /**
-     * Creates new form ClientForm
-     */
     public CustomerForm() {
         runView();
     }
 
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
     @SuppressWarnings("unchecked")
     // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
     private void initComponents() {
@@ -488,15 +476,7 @@ public final class CustomerForm extends javax.swing.JFrame implements ViewForm {
         }
     }//GEN-LAST:event_txtCostKeyTyped
 
-    /**
-     * @param args the command line arguments
-     */
     public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
         try {
             for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
                 if ("Nimbus".equals(info.getName())) {
@@ -507,8 +487,6 @@ public final class CustomerForm extends javax.swing.JFrame implements ViewForm {
         } catch (ClassNotFoundException | InstantiationException | IllegalAccessException | javax.swing.UnsupportedLookAndFeelException ex) {
             java.util.logging.Logger.getLogger(CustomerForm.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         }
-        
-        /* Create and display the form */
         java.awt.EventQueue.invokeLater(new Runnable() {
             @Override
             public void run() {
@@ -545,8 +523,6 @@ public final class CustomerForm extends javax.swing.JFrame implements ViewForm {
     public javax.swing.JTextField txtCost;
     public javax.swing.JTextField txtName;
     // End of variables declaration//GEN-END:variables
-
-    
 
     private void loadItemPlan() {
         cmbPlan.addItem("VISITA");
