+++ src/main/java/com/aztsoft/gym/view/forms/model/CustomerTableModel.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.aztsoft.gym.view;
+package com.aztsoft.gym.view.forms.model;
 
 import com.aztsoft.gym.domain.CustomerRegistration;
